pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fast-aiogentic"
version = "0.1.0"
description = "Fast-agent powered aiogram Telegram bot template"
readme = "README.md"
requires-python = ">=3.13.5,<3.14"
license = {text = "MIT"}
authors = [
    { name = "Jona Heidsick", email = "jona.heidsick@gmail.com" },
]
dependencies = [
    # Telegram bot framework (dev-3.x for aiohttp compat with fast-agent)
    "aiogram @ git+https://github.com/aiogram/aiogram@dev-3.x",
    # Core
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    # Fast-agent integration
    "fast-agent-mcp>=0.4",
    "telegramify-markdown>=0.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
    "ty>=0.0.1a7",
    "pre-commit>=3.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/fast_aiogentic"]

[tool.uv.sources]
aiogram = { git = "https://github.com/aiogram/aiogram", rev = "dev-3.x" }

# ============================================================================
# RUFF - 22 Rule Categories (Strict)
# ============================================================================
[tool.ruff]
line-length = 120
target-version = "py313"
src = ["src", "tests"]
exclude = [".git", ".venv", "build", "dist", "__pycache__"]

[tool.ruff.lint]
select = [
    "E", "W", "F",
    "I", "TID", "TCH",
    "BLE", "TRY", "EM", "RSE",
    "N", "UP", "B", "A",
    "C4", "C901", "SIM",
    "RUF", "PL", "PIE",
    "PERF", "FURB", "RET",
    "DTZ", "ANN401",
    "FIX", "ERA", "T10",
    "PTH", "LOG", "G", "PT",
]
ignore = [
    "PLR0913",  # Too many arguments (Pydantic)
    "PLR2004",  # Magic value comparison
    "TRY003",   # Long exception messages
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
known-first-party = ["fast_aiogentic"]
force-single-line = true

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pylint]
max-args = 8

# ============================================================================
# TY - Type Checking (80x faster than mypy)
# ============================================================================
[tool.ty.environment]
python-version = "3.13"

[tool.ty.rules]
unresolved-import = "error"
unresolved-reference = "error"
invalid-assignment = "error"

# Suppress false positives from pydantic-settings and fast-agent dynamic APIs
[[tool.ty.overrides]]
include = ["src/fast_aiogentic/config.py"]
[tool.ty.overrides.rules]
missing-argument = "ignore"

[[tool.ty.overrides]]
include = ["src/fast_aiogentic/agent_bridge.py"]
[tool.ty.overrides.rules]
invalid-assignment = "ignore"
unresolved-attribute = "ignore"

[[tool.ty.overrides]]
include = ["src/fast_aiogentic/handlers.py"]
[tool.ty.overrides.rules]
invalid-argument-type = "ignore"

# ============================================================================
# PYTEST
# ============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --strict-markers --cov=src/fast_aiogentic --cov-report=term-missing"

[tool.coverage.run]
branch = true
source = ["src/fast_aiogentic"]
omit = ["*/tests/*"]

[tool.coverage.report]
precision = 2
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
